% Generated by roxygen2 (4.0.2): do not edit by hand
\name{constant}
\alias{constant}
\title{A signal with a constant value.}
\usage{
constant(x)
}
\arguments{
\item{x}{A value}
}
\description{
This is a wrapper aroud \code{reactive()}, used to make it clear that
you really do want a constant.
}
\examples{
constant(10)
}

