% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count}
\alias{count}
\alias{countIf}
\title{Count the number of times a signal has been updated.}
\usage{
count(signal)

countIf(signal, predicate)
}
\arguments{
\item{signal}{A signal.}

\item{predicate}{A predicate; only signals for which the predicate
returns true will be counted.}
}
\value{
A reactive integer
}
\description{
Count the number of times a signal has been updated.
}
\examples{
rand <- reactive({fps(30); rnorm(1)})
count(rand)
countIf(rand, function(x) x > 0)
}

