% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filtering}
\alias{dropIf}
\alias{dropWhen}
\alias{filtering}
\alias{keepIf}
\alias{keepWhen}
\title{Filter signals.}
\usage{
keepIf(signal, predicate, init = NULL)

dropIf(signal, predicate, init = NULL)

keepWhen(signal, predicate, init = NULL)

dropWhen(signal, predicate, init = NULL)
}
\arguments{
\item{signal}{A signal}

\item{predicate}{For \code{if} functions, a predicate function;
for \code{when}, a signal yielding TRUE or FALSE.}

\item{init}{Shiny does not allow undefined signals, so an initial value
must be provided in case the first signal is filtered out.}
}
\description{
There are two variations: do you want to use a predicate function (\code{if})
or a boolean stream (\code{when}); and do you want to keep or drop the
values.
}

