% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fps}
\alias{fps}
\title{Fire a signal at fixed fps.}
\usage{
fps(rate)
}
\arguments{
\item{rate}{Number of signals to fire per second}
}
\value{
A reactive giving the time in seconds between the current and
  previous signal being fired
}
\description{
Fire a signal at fixed fps.
}
\examples{
\donttest{
shinyApp(
  ui = fluidPage(textOutput("delta")),
  server = function(input, output) {
    tick <- fps(10)
    output$delta <- renderText(tick())
  }
)
}
}

