% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fpsWhen}
\alias{fpsWhen}
\title{Conditionally file a signal at fixed fps.}
\usage{
fpsWhen(rate, when, session = NULL)
}
\arguments{
\item{rate}{Number of signals to fire per second}

\item{when}{A reactive that evaluates to TRUE or FALSE.}
}
\value{
A reactive giving the time in seconds since the last signal.
  The first time after a pause is always zero - this way summing the
  signal gives the amount of sum the output signal has been running
}
\description{
Conditionally file a signal at fixed fps.
}
\examples{
shinyApp(
  ui = fluidPage(textOutput("delta"), checkboxInput("pause", "pause")),
  server = function(input, output) {
    tick <- fpsWhen(10, reactive(!input$pause))
    output$delta <- renderText(tick())
  }
)
}

