% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lift}
\alias{\%~>\%}
\alias{lift}
\alias{lift2}
\title{Lift a function onto a reactive.}
\usage{
lift(signal, f)

lift2(signal1, signal2, f)

signal \%~>\% f
}
\arguments{
\item{signal,signal1,signal2}{Signals.}

\item{f}{A function to apply: for \code{lift()} it should take a single
argument, for \code{lift2()} it should take two.}
}
\description{
Lifting a function onto a signal creates a new signal that's been
transformed by the function. \code{\%~>\%} is an infix conveneience.
}

