% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge}
\alias{merge}
\title{Merge two signals.}
\usage{
merge(signal1, signal2)
}
\arguments{
\item{signal1,signal2}{Signals. The signals should return the same
type of object.}
}
\description{
Creates a new signal that contains the signals generated by both inputs.
}

