% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reducePast}
\alias{reducePast}
\title{Reduce a signal with its past values.}
\usage{
reducePast(signal, fun, init)
}
\arguments{
\item{signal}{A signal.}

\item{fun}{An accumulator function. Should take two arguments}

\item{init}{An initial value.}
}
\value{
A reactive of the same type as \code{init}.
}
\description{
Reduce a signal with its past values.
}
\examples{
# Simulate a markov random walk
walk <- reactive({fps(30); sample(c(-1, 0, 1), 1)})
location <- walk \%>\% reducePast(`+`, 0)

# Eventually:
# reactivePlot(count(walk), location)
}

