% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{since}
\alias{since}
\title{Since last sigal?}
\usage{
since(signal, time)
}
\arguments{
\item{signal}{A reactive.}

\item{time}{Delay in seconds.}
}
\value{
A reactive boolean.
}
\description{
Has it been \code{time} seconds since the last signal?
}
\examples{
\donttest{
shinyApp(
  ui = fluidPage(actionButton("click", "click"), textOutput("clicked")),
  server = function(input, output) {
    clicked <- reactive(input$click) \%>\% since(2)
    output$clicked <- renderText(clicked())
  }
)
}
}

