% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fpsWhen}
\alias{fpsWhen}
\title{Conditionally file a signal at fixed fps.}
\usage{
fpsWhen(when, rate)
}
\arguments{
\item{when}{A reactive that evaluates to TRUE or FALSE.}

\item{rate}{Number of signals to fire per second}
}
\value{
A reactive giving the time in seconds since the last signal.
  The first time after a pause is always zero - this way summing the
  signal gives the amount of sum the output signal has been running
}
\description{
Conditionally file a signal at fixed fps.
}
\examples{
\donttest{
shinyApp(
  ui = fluidPage(
    textOutput("delta"),
    textOutput("running"),
    textOutput("count"),
    checkboxInput("pause", "pause")
  ),
  server = function(input, output) {
    tick <- reactive(!input$pause) \%>\% fpsWhen(10)
    running <- tick \%>\% reducePast(`+`, 0)
    count <- tick \%>\% count()
    output$delta <- renderText(tick())
    output$running <- renderText(running())
    output$count <- renderText(count())
  }
)
}
}

