% Generated by roxygen2 (4.0.2): do not edit by hand
\name{since}
\alias{since}
\title{Since last sigal?}
\usage{
since(signal, delay)
}
\arguments{
\item{signal}{A reactive.}

\item{delay}{Delay in seconds.}
}
\value{
A reactive boolean.
}
\description{
Has it been \code{delay} seconds since the last signal?
}
\examples{
shinyApp(
  ui = fluidPage(actionButton("click", "click"), textOutput("clicked")),
  server = function(input, output) {
    clicked <- reactive(input$click) \%>\% since(2)
    output$clicked <- renderText(clicked())
  }
)
}

