% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timestamp}
\alias{timestamp}
\title{Add a timestamp to any signal.}
\usage{
timestamp(signal)
}
\arguments{
\item{signal}{Input signal}
}
\value{
A reactive list: the first element is a timestamp (a
  \code{\link{POSIXct}}), the second element is the \code{signal}.
}
\description{
Add a timestamp to any signal.
}
\examples{
\donttest{
shinyApp(
  ui = fluidPage(textOutput("tick"), textOutput("time")),
  server = function(input, output) {
    tick <- fps(10) \%>\% timestamp()
    output$tick <- renderText(tick()[[1]])
    output$time <- renderText(tick()[[2]])
  }
)
}
}

